agents/langgraph_app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
sqlalchemy==2.0.35
pymysql==1.1.1
pymongo==4.9.2
python-jose[cryptography]==3.3.0
cryptography==43.0.1
httpx==0.27.2
langchain==1.1.3
langgraph==1.0.4
langgraph-prebuilt==1.0.5
langchain-openai==1.0.3
